Add token scanning for single-char token types
Added the logic for building tokens out of single-char lexemes (e.g.
"+", "(", ")", etc.)

Added ERR_TOKENIZER_UNKNOWN_TOKEN error type. This error type is used
when the tokenizer finds a character that's not a part of the language
syntax in which case the tokenizer should abort tokenizing and return
the error.