Append tokens from scan_token
scan_token used to populate the token passed to it but not add it to the
tokens list, delegating that to the caller. This turned out to be a bad
idea since scan_token needs to return nothing when it encounters a
"skipable" piece of code (whitespace, newlines, comments, etc.). This
means that if I want to return something from scan_token, it has to be
Token* and, generally speaking, I don't want to deal with heap
allocations here.

This commit brings in a refactor that makes scan_token append the
generated token to the list if there is one, making it a little more
flexible and skipping the need for mallocs.