Setup specs and build strategies
I really needed a way of testing the code, however, I didn't want to
install a third-party library as it doesn't seem to be that necessary at
this point.

I decided to write specs via regular C functions but using a couple of
custom-built helpers to make them looks more like normal specs (these
are not fully implemented yet, of course).

When it comes to building, each "spec.c" file will be compiled into its
own executable. Alongside "spec.c", there should be a "build.cmake" file
that specifies spec-specific source files that need to be included into
the executable (spec helpers are included into each spec and do not need
to be listed in build.cmake). Both "spec.c" and "build.cmake" are files
that the CMakeLists expects to exist in the given dir when building a
spec executable. This makes it easier to reuse some CMake code and
reduce the amount of lines and frustration needed to add a new spec to
CMake.